[bp/r59396][bp/r59397][bp/r59395][MOD]Change wiki_page_name_above pref to wiki_page_name_position, and also made it a list rather than a checkbox, so we can have more title display options to choose from.[FIX] Hide default title when other option is chosen at wiki_page_name_position. [ENH] Adds an option for page_title_position to be inside page-data content.[Revert]Added the wiki_page_name_above back, to back port this to 15.x
This BP I did some changes comparing to the trunk, to keep the wiki_page_name_above pref working, to whom may use it. In trunk I replaced it with wiki_page_name_position, as there is no need to keep both. If you people think this is not needed, we can revert it, and merge the original from trunk. What you think?

git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/15.x@59407 b456876b-0849-0410-b77d-98878d47e9d5